+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,3 +1,4 @@
+import { User } from '../../types/User';
 import { clsx } from 'clsx';
 import { UserInfo } from '../UserInfo';
 
@@ -5,10 +6,10 @@ interface Props {
   id: number;
   title: string;
   completed: boolean;
-  userId: number;
+  user: User;
 }
 
-export const TodoInfo = ({ id, title, completed, userId }: Props) => {
+export const TodoInfo = ({ id, title, completed, user }: Props) => {
   return (
     <article
       data-id={id}
@@ -18,7 +19,7 @@ export const TodoInfo = ({ id, title, completed, userId }: Props) => {
     >
       <h2 className="TodoInfo__title">{title}</h2>
 
-      <UserInfo id={userId} />
+      <UserInfo user={user} />
     </article>
   );
 };
