+++ src/types/Todo.ts
@@ -1,6 +1,9 @@
+import { User } from './User';
+
 export interface Todo {
   id: number;
   title: string;
   completed: boolean;
   userId: number;
+  user: User;
 }
